requirements.txt
python=3.8
pandas
google-analytics-data
snowflake-snowpark-python